Write engine build info into the windows registry
